Pixmap and pixel format-related updates
* Imagine: Rewrite Pixmap and related classes for improved usability
* Imagine: Rewrite PixelFormat and related classes for improved usability
* Imagine: Clean up dynamic loading code for AMotionEvent_getAxisValue
* Imagine: Make string_printf a direct wrapper for snprintf and forward format string along with variable args
* Imagine: Add missing implementation of TimeIOS::makeWithSecsD
* EmuFramework: Don't store pixmap for on-screen d-pad bounding box or overlay effect

+++ imagine/src/gfx/opengl/GeomQuadMesh.cc
@@ -87,7 +87,7 @@ void GeomQuadMesh::setColorRGB(ColorComp r, ColorComp g, ColorComp b)
 {
 	iterateTimes(verts, i)
 	{
-		v[i].color = VertexColorPixelFormat.build(r, g, b, VertexColorPixelFormat.a(v[i].color));
+		v[i].color = VertexColorPixelFormat.build((uint)r, (uint)g, (uint)b, VertexColorPixelFormat.a(v[i].color));
 	}
 }
 
@@ -95,19 +95,19 @@ void GeomQuadMesh::setColorTranslucent(ColorComp a)
 {
 	iterateTimes(verts, i)
 	{
-		v[i].color = VertexColorPixelFormat.build(VertexColorPixelFormat.r(v[i].color), VertexColorPixelFormat.g(v[i].color), VertexColorPixelFormat.b(v[i].color), a);
+		v[i].color = VertexColorPixelFormat.build(VertexColorPixelFormat.r(v[i].color), VertexColorPixelFormat.g(v[i].color), VertexColorPixelFormat.b(v[i].color), (uint)a);
 	}
 }
 
 void GeomQuadMesh::setColorRGBV(ColorComp r, ColorComp g, ColorComp b, uint i)
 {
-	v[i].color = VertexColorPixelFormat.build(r, g, b, VertexColorPixelFormat.a(v[i].color));
+	v[i].color = VertexColorPixelFormat.build((uint)r, (uint)g, (uint)b, VertexColorPixelFormat.a(v[i].color));
 }
 
 void GeomQuadMesh::setColorTranslucentV(ColorComp a, uint i)
 {
 	// swap for tri strip
-	v[i].color = VertexColorPixelFormat.build(VertexColorPixelFormat.r(v[i].color), VertexColorPixelFormat.g(v[i].color), VertexColorPixelFormat.b(v[i].color), a);
+	v[i].color = VertexColorPixelFormat.build(VertexColorPixelFormat.r(v[i].color), VertexColorPixelFormat.g(v[i].color), VertexColorPixelFormat.b(v[i].color), (uint)a);
 }
 
 void GeomQuadMesh::setPos(GC x, GC y, GC x2, GC y2)
